Move global types declaration and new types for router.
