+++ src/reducers/chatReducer.ts
@@ -1,5 +1,5 @@
 import {inferActionsType} from '../redux/reduxStore';
-import {baseThunkType} from './types/types';
+import {baseThunkType} from '../types';
 import {chatApi, messageType} from '../components/API/chat-api';
 import {Dispatch} from 'redux';
